requirements.txt
Django==1.8.12
mysqlclient==1.3.3
Pillow==3.2.0
django-colorful==1.2
